Restructure: move files into website/ package, now includes __init__.py
