+++ restructured python package to include __init__.py
@@ -1,0 +1,21 @@
+
+# Please enter the commit message for your changes. Lines starting
+# with '#' will be ignored, and an empty message aborts the commit.
+#
+# On branch restructure
+# Your branch is up to date with 'origin/restructure'.
+#
+# Changes to be committed:
+#	modified:   app.py
+#	deleted:    routes.py
+#	new file:   website/__init__.py
+#	new file:   website/__pycache__/__init__.cpython-312.pyc
+#	new file:   website/__pycache__/routes.cpython-312.pyc
+#	renamed:    models.py -> website/models.py
+#	new file:   website/routes.py
+#	renamed:    static/css/styles.css -> website/static/css/styles.css
+#	renamed:    static/js/script.js -> website/static/js/script.js
+#	renamed:    templates/base.html -> website/templates/base.html
+#	renamed:    templates/dashboard.html -> website/templates/dashboard.html
+#	renamed:    templates/index.html -> website/templates/index.html
+#
